@@ -1,1 +1,1 @@
-ruby-1.9.2-p290
+ruby-2.2.1
