@@ -1,3 +1,3 @@
 module BlueJay
-	VERSION = "3.5"
+	VERSION = "4.0"
 end
